Pptx added

+++ src/ISP Desk/Model/Service.cs
@@ -4,6 +4,6 @@
     {
         public int ServiceID { get; set; }
         public string ServiceName { get; set; }
-        public int AdditionalPrice { get; set; }
+        public decimal AdditionalPrice { get; set; }
     }
 }
