+++ src/ISP Desk/Model/Tariff.cs
@@ -4,7 +4,7 @@
     {
         public int TariffID { get; set; }
         public string TariffName { get; set; }
-        public int TariffPrice { get; set; }
+        public decimal TariffPrice { get; set; }
         public List<Service> Services { get; set; }
     }
 }
